Extracet Compiler class
